pyproject.toml
[tool.poetry]
name = "qap-sandbox"
version = "0.1.0"
description = ""
authors = ["Mikhail Lyalin <mc.lyalin@gmail.com>"]
readme = "README.md"
packages = [{include = "qap_sandbox"}]

[tool.poetry.dependencies]
python = "^3.11"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
pytest-selenium = "^4.0.0"
webdriver-manager = "^3.8.5"
black = "^22.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
